+++ squads/src/App.css
@@ -91,7 +91,7 @@ p{
 		align-items: center;
 		justify-content: center;
 		width: 100%;
-		height: 100vh;
+		height: 70vh;
 		text-align: center;
 		scroll-snap-align: center;
